feat(sessions): IPC commands, state broadcasts, dock deep link, needs-input notification
Seven thin commands over the pool; SessionsChanged/SessionStateChanged
specta events wired into the pool's transition path; katto://dock route
for the needs-input notification click.

+++ src-tauri/src/sessions.rs
@@ -45,5 +45,5 @@ pub struct SessionInfo {
     pub state: SessionState,
     pub cwd: String,
     pub started_at: String,
-    pub idle_since_secs: Option<u64>,
+    pub idle_since_secs: Option<u32>,
 }
